feat. 찜관련 기능 구현중

+++ src/main/java/com/ssginc/unnie/shop/dto/ShopBookmarkRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class ShopBookmarkRequest {
-    private int bookmarkMemberId;
+    private long bookmarkMemberId;
     private int bookmarkShopId;
 }
